requirements.txt
torch == 1.3.1
tqdm >= 4.35.0
pandas >= 0.24.2
matplotlib >= 3.1.1
h5py >= 2.9.0
scikit - learn >= 0.22
seaborn >= 0.9.0
torchvision == 0.4.0
pillow >= 7.2.0
opencv - python >= 3.4.8.29
scikit - image >= 0.17.2
dgl == 0.4.3.post2
PyYAML >= 5.1.2
